Refactor grain size evolution in separate reaction model
